chore: refactor ui into separate components
